Refactor: MarketIndexSummary 기능으로 분리
