+++ src/features/marketIndexSummary/index.ts
@@ -1,0 +1,1 @@
+export { MarketIndexSummary } from "./ui/MarketIndexSummary";
